HBASE-13329 ArrayIndexOutOfBoundsException in CellComparator#getMinimumMidpointArray.

+++ hbase-common/src/main/java/org/apache/hadoop/hbase/CellComparator.java
@@ -472,7 +472,7 @@ public class CellComparator implements Comparator<Cell>, Serializable {
       final byte [] rightArray, final int rightOffset, final int rightLength) {
     // rows are different
     int minLength = leftLength < rightLength ? leftLength : rightLength;
-    short diffIdx = 0;
+    int diffIdx = 0;
     while (diffIdx < minLength &&
         leftArray[leftOffset + diffIdx] == rightArray[rightOffset + diffIdx]) {
       diffIdx++;
